Add data, WorkoutLinkTag component
Created a data folder to keep the workouts.json file isolated, added JS
ON file to the new data folder. Created a WorkoutLinkTag component to s
erve as the clickable link component which a user would use to access c
ontent. Updated Dashboard page to import the new JSON data, and map ove
r it console logging the data. Object import appears to be successful,a
nd a log is created for each item in the workout JSON file.