Add Global CG CSS Lib, add Header component
Added a Header component to App.js. Will serve as page branding by disp
laying the CG logo, as well as any navigational links if needed. Added
CDN to index.html, but unfortunately was not connecting as expected. In
stead I went ahead and took the minified CSS file and added it to the i
ndex.css file. Logo and styling now appear to be applying correctly.