Update WorkoutLinkTag onClick navigation
Updated the WorkoutLinkTag component to listen for a click event on the
containing div element. Also converted the WorkoutLinkTag component to
import withRouter from RRD, and pass the entire component to it as argu
ment. This will return a HOC which has access to the browser history AP
I and allows the onClick to push a new URL to the history stack, changi
ng the page to the VideoWorkout component page. All tags currently push
to the correct VideoWorkout component, but will need to update content
to render dynamically based on props and specified trainers/videos.