Update workouts route, WorkoutLinkTag styles
Updated the route in App.js for 'workouts' to be '/workouts/:workoutId'
which will allow the URL to be dynamically rendered based on what is su
bstituted for the :workoutId parameter. Also updated WorkoutLinkTag com
ponent to push on user click the value '/workouts/' followed by the tra
inerId. This will allow the VideoWorkouts page to render the appropriat
e videos based on that ID. Clicking on the Tag currently links to the c
orrect URL, but VideoWorkouts does not yet populate that data.