Render content in VideoWorkouts, add styles
Updated VideoWorkouts page to dynamically render workout content using
the withRouter HOC, and the match property. Currently it will assign th
e URL parameter of workoutId to a constant called trainerId, which was
sent from the Dashboard WorkoutLinkTag push and will be an integer. Thi
s number is used to access a particular object in the workouts.json fil
e, giving access to the media and other properties within. Based on the
se properties, the title and the video stream can be rendered.Appears t
o be working as expected for each button on the Dashboard page.