Update dashboard stylings
Further updating Dashboard page styles so as to better align with CG st
andards for design patterns per UI library docs. Replaced a div with a
section tag so as to increase semantic value and accessibility. Added c
lass on heading to use CG css library as opposed to external CSS page.
Lastly, added comments for ease of discussion in presentation and clari
ty to readers.